move field type -> es mapping to packages/events
