+++ packages/events/src/service/indexing/indexing.ts
@@ -10,12 +10,13 @@
  */
 
 import {
+  AddressFieldValue,
   EventConfig,
   EventDocument,
   EventIndex,
   FieldConfig,
-  getCurrentEventState,
-  mapFieldTypeToElasticsearch
+  FieldType,
+  getCurrentEventState
 } from '@opencrvs/commons/events'
 import { type estypes } from '@elastic/elasticsearch'
 import * as eventsDb from '@events/storage/mongodb/events'
@@ -118,6 +119,62 @@ export function encodeFieldId(fieldId: string) {
 
 function decodeFieldId(fieldId: string) {
   return fieldId.replaceAll(SEPARATOR, '.')
+}
+
+function mapFieldTypeToElasticsearch(field: FieldConfig) {
+  switch (field.type) {
+    case FieldType.DATE:
+      // @TODO: This should be changed back to 'date'
+      // When we have proper validation of custom fields.
+      return { type: 'text' }
+    case FieldType.TEXT:
+    case FieldType.PARAGRAPH:
+    case FieldType.BULLET_LIST:
+    case FieldType.PAGE_HEADER:
+      return { type: 'text' }
+    case FieldType.DIVIDER:
+    case FieldType.RADIO_GROUP:
+    case FieldType.SELECT:
+    case FieldType.COUNTRY:
+    case FieldType.CHECKBOX:
+    case FieldType.LOCATION:
+      return { type: 'keyword' }
+    case FieldType.ADDRESS:
+      const addressProperties = {
+        country: { type: 'keyword' },
+        province: { type: 'keyword' },
+        district: { type: 'keyword' },
+        urbanOrRural: { type: 'keyword' },
+        town: { type: 'keyword' },
+        residentialArea: { type: 'keyword' },
+        street: { type: 'keyword' },
+        number: { type: 'keyword' },
+        zipCode: { type: 'keyword' },
+        village: { type: 'keyword' }
+      } satisfies {
+        [K in keyof Required<
+          NonNullable<AddressFieldValue>
+        >]: estypes.MappingProperty
+      }
+      return {
+        type: 'object',
+        properties: addressProperties
+      }
+    case FieldType.FILE:
+      return {
+        type: 'object',
+        properties: {
+          filename: { type: 'keyword' },
+          originalFilename: { type: 'keyword' },
+          type: { type: 'keyword' }
+        }
+      }
+    default:
+      const _exhaustiveCheck: never = field
+      throw new Error(
+        `Unhandled field type: ${JSON.stringify(_exhaustiveCheck)}`
+      )
+  }
 }
 
 function formFieldsToDataMapping(fields: FieldConfig[]) {
